requirements.txt
functions-framework==3.*
google-generativeai==0.5.0